🔧 chore: Update schema
