+++ docs/changelog/index.json
@@ -2,18 +2,51 @@
   "$schema": "https://github.com/lobehub/lobe-chat/blob/main/docs/changelog/schema.json",
   "cloud": [
     {
-      "id": "cloud-market",
-      "versionRange": ["1.35.1", "1.2.1"]
+      "id": "agent-market",
+      "date": "2024-11-10",
+      "versionRange": ["1.35.1", "1.33.1"]
     }
   ],
   "community": [
     {
       "id": "agent-market",
-      "versionRange": ["1.0.0", "1.0.1"]
+      "date": "2024-11-20",
+      "versionRange": ["1.33.0", "1.20.1"]
     },
     {
-      "id": "cloud-market",
-      "versionRange": ["1.35.0", "1.2.1"]
+      "id": "agent-market",
+      "date": "2024-11-12",
+      "versionRange": ["1.20.0", "1.12.1"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-11-12",
+      "versionRange": ["1.2.1"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-10-12",
+      "versionRange": ["0.163.0", "0.143.0"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-08-02",
+      "versionRange": ["0.143.0"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-07-02",
+      "versionRange": ["0.143.0"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-06-02",
+      "versionRange": ["0.143.0"]
+    },
+    {
+      "id": "agent-market",
+      "date": "2024-05-12",
+      "versionRange": ["0.143.0"]
     }
   ]
 }
